split dataframe.py into multiple files
